requirements.txt
matplotlib==3.10.1
numpy==2.2.4
streamlit==1.43.2
requests>=2.31
python-dotenv>=1.0,<2.0